requirements.txt
anyio @ file:///C:/ci/anyio_1644481856696/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569876605/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1643036627027/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clr-loader==0.2.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1643209086516/work
cycler==0.11.0
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dwf==0.1.0
entrypoints==0.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.30.0
future==0.18.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.16.1
importlib-metadata @ file:///C:/ci_310/importlib-metadata_1642082991434/work
ipykernel @ file:///C:/ci/ipykernel_1647000977835/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1646948607209/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
iso8601==1.0.2
jedi @ file:///C:/ci/jedi_1644315428305/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1646994639510/work
jupyter-server @ file:///opt/conda/conda-bld/jupyter_server_1644494914632/work
jupyterlab @ file:///opt/conda/conda-bld/jupyterlab_1644830542042/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///opt/conda/conda-bld/jupyterlab_server_1644500396812/work
kiwisolver==1.3.2
MarkupSafe @ file:///C:/ci_310/markupsafe_1642084045935/work
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///C:/ci_310/mistune_1642084168466/work
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient @ file:///tmp/build/80754af9/nbclient_1645431659072/work
nbconvert @ file:///C:/ci_310/nbconvert_1642173634377/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.7.1
notebook @ file:///C:/ci/notebook_1644984554467/work
numpy==1.22.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pco==0.1.3
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci_310/pyrsistent_1642117077485/work
pyserial==3.5
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pythonnet==3.0.1
pytz==2021.3
PyWavelets==1.3.0
pywin32==302
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp310-none-win_amd64.whl
PyYAML==6.0
pyzmq @ file:///C:/ci_310/pyzmq_1642090360481/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-image==0.19.2
scipy==1.8.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/ci_310/sniffio_1642092172680/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
terminado @ file:///C:/ci/terminado_1644312135040/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tifffile==2022.2.9
tk==0.1.0
tornado @ file:///C:/ci_310/tornado_1642093111997/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///C:/ci_310/websocket-client_1642093970919/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
